requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.28
alembic>=1.13.1
psycopg2-binary>=2.9.9
pymysql>=1.1.0
cryptography>=42.0.5
pydantic-settings>=2.2.1
google-genai>=0.3.0
pyyaml>=6.0.1
sse-starlette>=2.0.0
pytest>=8.1.1
pytest-asyncio>=0.23.6
